requirements.txt
fastapi
uvicorn
pydantic
requests
langchain
langchain-google-genai
google-generativeai
python-dotenv
langchain-community
uuid